+++ javatools/src/main/java/org/xvm/asm/constants/Float32Constant.java
@@ -67,7 +67,7 @@ public class Float32Constant
      */
     @Override
     public Float getValue() {
-        return Float.valueOf(m_flVal);
+        return m_flVal;
     }
